Extract output value resolvers from OutputRenderer
Begin splitting the 1251-line OutputRenderer by moving the cohesive, pure
sketch-document and timeline-sequence shape resolvers (isRecord, get*Id,
is*Like, unwrap*Candidate, resolve*) into outputValueResolvers.ts and
importing them. Behavior-preserving; OutputRenderer tests pass and typecheck
is clean. Type-specific renderers can follow into a registry next.
